feat: expose RwLockWriteGuard to sysvar_cache field of TransactionProcessor
Sometimes it's necessary to have mutable access to inner SysvarCache
field of TransactionProcessor, which has been made completely private,
only exposing a shared access via acquiring read lock on it.

This PR adds extra method to obtain a write lock on this field, for use
cases when mutable access is required.